requirements.txt
langchain==0.1.9 
langchain-community==0.0.21 
langchain-core==0.1.52 
python-dotenv==0.19.0   
google-search-results==2.4.2
streamlit
crewai==0.1.30
openai==1.6.1
langchain==0.0.352
tavily-python==0.2.6